# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PoupaMais - Personal Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: List all financial goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a financial goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a financial goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /goals/{id}/contribute:
    post:
      operationId: contributeToGoal
      tags: [goals]
      summary: Add contribution to a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalContribution"
      responses:
        "200":
          description: Contribution added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get monthly financial summary
      parameters:
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [dashboard]
      summary: Get spending breakdown by category
      parameters:
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success
...
    Goal:
      type: object
      required: [id, userId, name, targetAmount, currentAmount, type, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        type:
          type: string
          enum: [savings, travel, emergency, purchase, other]
        deadline:
          type: ["string", "null"]
        monthlyContribution:
          type: ["number", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        isCompleted:
          type: boolean
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [name, targetAmount, type]
      properties:
        name:
          type: string
          minLength: 1
        targetAmount:
          type: number
          minimum: 1
        currentAmount:
          type: ["number", "null"]
        type:
          type: string
          enum: [savings, travel, emergency, purchase, other]
        deadline:
          type: ["string", "null"]
        monthlyContribution:
          type: ["number", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    GoalUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        targetAmount:
          type: ["number", "null"]
        currentAmount:
          type: ["number", "null"]
        type:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        monthlyContribution:
          type: ["number", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    GoalContribution:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 0.01
...
    DashboardSummary:
      type: object
      required: [totalBalance, monthlyIncome, monthlyExpenses, monthlySavings, savingsRate, transactionCount]
      properties:
        totalBalance:
          type: number
        monthlyIncome:
          type: number
        monthlyExpenses:
          type: number
        monthlySavings:
          type: number
        savingsRate:
          type: number
        transactionCount:
          type: integer
        previousMonthExpenses:
          type: ["number", "null"]
        previousMonthIncome:
          type: ["number", "null"]
    CategorySpending:
      type: object
      required: [categoryId, categoryName, categoryColor, categoryIcon, amount, percentage, transactionCount]
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: string
        categoryIcon:
          type: string
        amount:
          type: number
        percentage:
          type: number
        transactionCount:
          type: integer
...