# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PoupaMais - Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalBalance: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  monthlySavings: number;
  savingsRate: number;
  transactionCount: number;
  /** @nullable */
  previousMonthExpenses?: number | null;
  /** @nullable */
  previousMonthIncome?: number | null;
}